perf(thermocycler): use timer interrupt for thermistor read and PID compute (#71)
adds a 25ms timer interrupt that updates a pair of thermistors on each interrupt. the loop cycles
through the next pair on each interrupt and the PID corresponding to that pair is computed right
after. This ensures that the compute occurs every 100ms. Changed PID library to add a sampling mode
selection. Switched ADC to 250 SPS

Closes #3610

* add comments, cleanup
* update readme, use better PID constants