temperature offset test script
The test script does offset testing with the current settings as follows: 95C to 70C each cycling ~32 seconds for 10 cycles.
This test script uses subprocess to execute data collection for the Eutechnics thermometers. The subprocess run independent from the test script. The reason for this is to get more sync data from all the devices.