chore(build): Switch to cmake as a configuration system (#125)
This PR switches away from raw makefiles as a configuration and build system and uses cmake instead.
Why?

We're about to enter territory that would be remarkably painful to tread upon with make: for upcoming module projects, we'll compile both for the target (probably with gcc) and the host (probably with clang); we'll also use other tools like clang sanitizers. We'll build in testing, which means - for native code projects - compiling different executables. All of this is hard to do in make, and the ways that you work around it typically mean abandoning the places where Make can help you the most.

Raw makefiles are pretty good at replacing shell scripts as a build system. They give you dependency management, they give you some built-ins for managing compiler options for compiler options. The farther you go from using them as a replacement for shell scripts, the worse they get. In particular, they really start to drag when the configuration of a build becomes more complex, and especially when there are multiple complex configurations. The typical way to avoid this in raw Make is to start avoiding using Make's builtin rules (so you don't have to, e.g., switch around $CC) or have increasingly complex switches around environment variables.

The other route that people take is to use a config system. There are a lot of these: kconfig, buildroot's config which somehow also runs under make in bizarre ways, IDE project configuration (which therefore ties you to an IDE), others.

But there's also cmake. Cmake is an ugly (just check out the code!) configuration system to look at, and its overeager attempts to help you out can occasionally get in your way, but it is fundamentally a solid, well-designed, extremely well-documented opinionated multi-platform native code configuration system that is widely used, integrates well with IDEs, and fixes the issues I noted above.
OK but why change what works?

Well, unfortunately, cmake's core desire is to compute a bunch of configuration and emit... Makefiles. Which means it's very hard to use alongside a different Makefile based system. Also I needed the practice, quite honestly.
So what changed?

opentrons-modules in this pr uses cmake as a configuration system, generating (currently) makefiles.

That means that the logic that used to be in the top-level Makefile is now spread out in the CMakeLists.txt files that exist in the source tree. Each file is responsible for building what is below it, in the parallel builds directory.
How do I use this?

(Check out the README)

The biggest usability downside of cmake is that it has two steps: configure, and build. You can do them both with the cmake command, which is your friend. The specific cmake you want to use is CMake 3.19, which you can find here. We want 3.19, which is the latest release, because of the new "presets" concept which allows much easier (MUCH easier) configuration of things like cross-compilers.

So first you configure: cmake .
And then you build: cmake --build ./builds

You can also use the generated Makefiles to build - cd builds && make.

To bundle up all the modules into a specific package, run cmake --install ./builds.

When you run the configuration step, cmake will automatically download the arduino IDE and configure it for you.