Add clang-tidy
Clang-tidy is a static analyzer for context-free problem discovery and
semantic-level style (e.g., "use -> return type" or "this line is
probably a bug"), as opposed to either semantically-meaningless code
style formatting (which is clang-format) or program-flow static
analysis (CodeCheck). It has its own config file, and can be run on
source using a compilation database, which it now is.