chore(stm32): add linters, checkers, tests (#127)
Adds clang-tidy, clang-format, and tests.

Also exports a compile-commands database.

Automatically downloads clang and associated tooling, if required (local installs preferred). In addition, host-gcc is used to compile tests.

Clang-tidy is a static analyzer for context-free problem discovery and
semantic-level style (e.g., "use -> return type" or "this line is
probably a bug"), as opposed to either semantically-meaningless code
style formatting (which is clang-format) or program-flow static
analysis (CodeCheck). It has its own config file, and can be run on
source using a compilation database, which it now is.


Co-authored-by: Alise Au <20424172+ahiuchingau@users.noreply.github.com>