feat(heater-shaker): FreeRTOS integration and tasks setup (#129)
Compiles in the FreeRTOS implementation build in to the STM32F303
BSP (not the CMSIS RTOS, though, because we are not ever going to change
the RTOS thank you very much and support and docs for FreeRTOS are much
more widely available).

This uses FreeRTOS's static allocation setup, in addition to (at this
point) heap implementation 5 - might change that. We want static
allocation for as much as we can get away with, and separate
per-use-case memory pools for the rest (e.g. for messages sent over
queues).

This also sets up and runs the main tasks that the firmware will have -
motor control, heater control, UI control, and host communications
control. They don't do anything yet though.

Also, move to github actions from travis.