feat(heater-shaker): add generic gcode parser
Add a generic (limited, single-level) parser-combinator style gcode
parser using templates.

This parser (in gcode_parser.hpp) parses gcodes, one at a time, out of a
"complete" gcode command string - a command string that ends in a
newline delimiter and therefore represents upstream's intent that the
command is complete.

While the parser isn't yet instantiated in the application code, because
there's not really anything that can feed it data yet, it is fully
tested and ready for use.