refactor(build): Make the bsp package have components
This is a lot cleaner since there's more or less 3 separate
subcomponents of the BSP (well, more, but for now we're only using 3,
and this gives a structure for adding more if we want to): the drivers
and the HAL, the USB middleware, and the FreeRTOS middleware. These are
nominally separate things, or at least
- FreeRTOS could theoretically be used without the drivers/hal (aside
from the device header, anyway)
- The drivers and HAL could certainly be used without either USB or
FreeRTOS
- USB support really needs the hal, no matter the documentation's
assertions to the contrary, but it could certainly be used without
freertos

The heater-shaker right now wants to use all three, and probably future
stm32 projects will as well, but separating these into components is
absolutely cleaner.