feat(heater-shaker): USB bringup (#131)
Adds USB support using the ST USB_Device middleware, presenting a CDC device class for serial emulation. As of this PR, the only implemented functionality is automatic loopback; it's not integrated into the rest of the firmware.

In addition,

- Make the bsp package have components

This is a lot cleaner since there's more or less 3 separate
subcomponents of the BSP (well, more, but for now we're only using 3,
and this gives a structure for adding more if we want to): the drivers
and the HAL, the USB middleware, and the FreeRTOS middleware. These are
nominally separate things, or at least
- FreeRTOS could theoretically be used without the drivers/hal (aside
from the device header, anyway)
- The drivers and HAL could certainly be used without either USB or
FreeRTOS
- USB support really needs the hal, no matter the documentation's
assertions to the contrary, but it could certainly be used without
freertos

The heater-shaker right now wants to use all three, and probably future
stm32 projects will as well, but separating these into components is
absolutely cleaner.

- gdb: add svd load support

This adds a (fairly shady) python gdb extension that can parse the cmsis
hardware description files provided by arm, and adds a command called
`svd` that lets you easily print the values of peripheral registers.
This requires some truly ugly work to download the couple of python
dependencies required and set up gdb's truly ugly python support, which
somehow ignores all system level hints about which python it should use
and dives straight at python 2.7 which doesn't have things like "pip
guaranteed to be installed" or "setuptools guaranteed to be
installed" (you have to install setuptools manually still sorry).