refactor(heater-shaker): queues: better recv
We had a try-receive that fails if there's nothing in the queue but
that's not actually what we want to do in tasks - we really do want to
block forever waiting for input, because then the task is asleep. We can
use try-receive if we ever need to add some periodic wakeups to check
other stuff, but that's for the future.

This is a bit problematic in a testing context because you don't want to
accidentally block forever if you call recv() before you've actually put
something in the backing deque, so throw if that happens.

We also need the ability to set timeouts in try_recv, since these
timeouts will drive the fixed-time frequencies for the controllers.