feat(heater-shaker): add ack cache
The ack cache is a layer required for driving a full through-system
acknowledge, where an ack is only returned to the host when the relevant
subsystem has truly implemented the command. We need to keep around the
gcode we'll be acking (so we can say so in our text interface) and we
need to associate it with an id (so we can pack it into the internal
messages) and therefore we need a place to do that associating. Normally
that would be a map but a map likes to allocate, and we don't want to do
that. We could either do a custom allocator, or have a flat array and
pay a linear search price that won't matter because we'll have like 8 entries.

ack cache formatting