feat(heater-shaker): integrate host comms with USB
Once we can communicate via usb, and we have a host comms task, we
should link the two, and let the task read and write. That means

- the USB subsystem has to be able to give the task data. The right way
to do this is to put it in the task queue, and the right way to do that
is with the freertos method to put stuff in a queue from an isr, which
is where we'll be in the CDC callback hook. So we can poke into our file
global to get a handle to the queue and do so. Of course, we need to
give the task time to handle the incoming message and not have a naughty
host blow through it with another USB transaction, so we have a nice
double buffer.

- the task has to be able to give the USB subsystem data. By structuring
the task's control flow as a function that runs through the task logic
one time, we can pass in a pointer and size to a tx double buffer, and
let the task logic fill it out; then we can hand it over to the usb
subsystem. This is also nice and testable.

- the task has to be able to actually handle and parse the input. The
task is receiving messages, in a std::variant, from its queue; we can
std::visit them. When the message is the one that the USB subsystem
sends, we can parse it with our gcode parser, and then std::visit the
parsed gcode. By passing down the buffer and length each time, we can
let the gcode parsing write back parse errors, or internal errors like
"oops destination queue full". By having those errors defined in a
header, we can write down what errors mean what.

- the task has to be able to handle the happy path of getting a message,
parsing a gcode, sending a message to the right subsystem, waiting for
that subsystem to handle it, then sending back an ack to the host. So we
need a table to keep the in-flight gcodes around.

All this gets managed in the host_comms_task!

There are also a couple usb fixes, such as actually giving the relevant
endpoints their own PMA buffer addresses so we can do things other than
loopback, and removing some very annoying entries in the default CDC
class descriptor that would cause linux kernels to try and control the
flow control lines and yell when it failed to do so.