fix(heater-shaker): task aggregator off stack
It turns out that when you start freertos it initializes its stack, and
while it's fine to have stuff on the stack _in tasks_, it will
absolutely blow away anything that was put on the stack by main(). The
only such thing that mattered was the task aggregator that gives all the
tasks references to the other tasks, so it had to move off the stack to
the free store. In turn, that means that it will be initialized before
the tasks are available, so it needs to hold pointers to tasks rather
than references to tasks; which then means that all the tasks need to
change, one way or another.