feat(heater-shaker): Add stm mcl to heater-shaker
This adds the ST motor control library to the build without building it
or integrating it in any way.

The ST motor control library consists of 3 parts:

- Precompiled static libraries specific to device and compiler
containing ST's special sauce for running the calculations themselves in
an efficient way (in MCLib/Lib)
- Library source and header files for the ancillary tasks that surround
the libraries and provide interfaces for them (MCLib/Any and MCLib/F3xx)
The above are static and copied out from the internals of the motor
control configurator but there's also...
- Autogenerated configuration and control code made by the STM32CubeMX
integration for the motor control library and the project file for that
configurator.

This last is the problematic one. The project file is an XML file that
really can't be subjected to text-based source control like git because
merging it will go really badly. Similarly, if the generator is actually
used beyond this point, it will generate a huge amount of noise in the
commit log. The autogenerated files will also be blown away and
regenerated except for code that is in the /** USER CODE **/-delineated
sections of the source file. It's not great!

So what we're going to do is not use it beyond the first-pass
configuration. The code in this commit is the direct output of that
generation pass; the project file is added too just in case. But beyond
this point, all this code will just be edited normally, and it's on
whoever uses that project file and regenerates the code to make sure
nothing gets broken.