feat(heater-shaker): integrate stm mcl into build
This commit integrates the motor control library into the build. This
requires
- Cmake changes to both build the library code itself and add the
integration files
- code changes in the integration files to adapt them to our tasks and
remove unused parts entirely (such as the UI side, which we won't be
supporting)

The code here is integrated in that it successfully builds and runs; it
does not actually do anything or control a motor.