feat(heater-shaker): Add motor policy
This will be most of the downward interface for the motor task. It will
be implemented differently in the firmware (where it will adapt the
STSpin api) and in the test (where it will provide mocking). The
implementation for STSpin is added here.

convert to/from rpm and 01hz in policy