feat(heater-shaker): send motor errors upstream
The motor control library will often see errors and log them in its
internals; we need to send them up to the user.

This adds a new message for the other internal motor tasks to send to
the main motor task, with a C++y definition for the error bitfields so
we don't have to include the firmware stuff upstream. There's also a
helper function to transform it into the new errors.

Finally, the error code got a bit of a rework to not have the strings
just live in errors.hpp and also not violate the one-definition rule.

We also need to adapt the motor testing to use the new polcies, so
there's a test motor policy that cna be used in the testing.

return stuff