feat(heater-shaker): Configure for softpower
The soft power button is the first element of the UI that we'll support.
It consists of, at this stage, making sure that the pin connected to the
button sense isn't pulled down because that breaks the softpower
circuit. It (and the pin for sensing the power being unplugged, though
that's less critical) needs to be an input without pull resistors.