chore(heater-shaker): add flash target
We have a target for debugging, but it's actually kind of annoying to
use it to do a flash where you don't actually want to debug.

Add a cmake target that just calls openocd to flash the board and exits.

Also add a couple comments (which get printed in the console) to some
other cmake utilities.