feat(heater-shaker): generate ntc lookup tables
The way you convert from NTC resistance (which is temperature dependent)
into a temperature is by using a vendor-provided lookup table (since
otherwise there's lots of lovely ln and exp in there that you don't want
to deal with). This is (in this case) provided in a csv; we'd like it as
c++ source. So let's use cmake's ability to add custom targets and use
the output of those custom targets really easily to check in the csv and
generate c++ source and headers using a python script.