feat(heater-shaker): add heater adc hardware bringup
The new heater_hardware follows or perhaps starts or intensifies the
pattern in the motor where there needs to be a bridge, written by us,
between the c code with c expectations in the hal and the c++ code with
c++ expectations in the rest of the firmware. For instance, just because
we can turn off diagnostics about improper (for c++) use of volatile,
that doesn't mean we can actually use the macros that cause the warning;
we need to do that from C.

So, heater_hardware.c and .h provide a C interface to the hardware that
is usable by the c++ side.

This code brings up the ADC used by the thermistor pins. It's configured
so that the C++ code kicks off a three-channel conversion sequence, with
an interrupt firing after each conversion and eventually (when all three
channels have been converted) firing off a callback that should be
provided by the c++ side.

There's an annoying bit of statefulness in this code around remembering
which adc channel is currently being converted; the hardware doesn't do
it, the hal doesn't do it, so we have to do it. This also makes the
defensive coding in the conversion complete callback a little tricky
because even if we're not initialized yet (not that the callback should
fire before initialization) we need to make sure we don't get out of
sync with the ADC's internal state.