feat(heater-shaker): thermistor conversion
This thermistor conversion class ingests ADC readings directly (with
parametrization about the adc bit depth and the hardware parameters for
the thermistor circuit) and uses the autogenerated tables for the
thermistor and some other math to convert the values to C. It's also
almost monadic but I'm too much of a coward to go all the way. The
Result types get carried through and bring any errors in the
intermediates back to the caller rather than having to try and signal it
in-band by carving out like -1 or -inf as an error signal.

The numbers on the bias resistance in the test are slightly weird
because I forgot that number needed to be in kOhm rather than Ohm but
the numbers in the tests aree all arbitrary anyway, it's the boundary
checking around the std::find that needs to be tested.