feat(heater-shaker): handle temperature updates
Add a message that will come from a hardware task and carry updated ADC
readings to the heater task. The heater task then converts the readings
to temperatures and handles any errors that result.

There needs to be a followup that handles things like the case where the
thermistor is broken and only gives bad readings; right now, we'll just
send an error every single time which isn't the most helpful thing in
the world.

As a sidenote, the temperature update messages going in provide a useful
but somewhat sideways way of specifying the control loop frequency. I
think it's ok that it's implicit since the actual sequence will be
specified by the value defined in this class.