feat(heater-shaker): integrate thermistor reading
As with the motor task, we need another internal task to drive timing
and hardware-dependent stuff we don't want to sully the perfection of
the main c++ task. The task in this case does a bit less than in the
motor; it just serves to kick off a temperature update every so often.
This happens every so often rather than as fast as possible for two
reasons:
- The physical time scale we're talking about here is thermal, which
means slow. There's no point reading these ADCs in the megahertz (or
kilohertz, or frankly hertz) range because temperatures don't change
that fast; NTCs don't react that fast; and we have an antialiasing
filter in front of us anyway.
- There's no point having a million interrupts firing all the time if
we're not going to use the values.

So instead, we have this heater hardware task that runs (for now) at
10Hz (which can probably be lowered to more like 1 or 0.5Hz) and kicks
off an ADC conversion, which will cause 3 interrupts, one for each
channel, and send a message to wake the heater task on the third. That
means the main heater task will also run at 10Hz, just 3 adc conversions later.