fix(heater-shaker): trigger conversions in isr
Annoyingly, this micro's ADCs can be set up in chain conversion mode but
they only give you an interrupt after all the conversions are complete,
so if you're not either polling to catch it or using DMA you won't see
the first two. Of course, we didn't have it in that mode anyway before
that, but that's why the fix is to start the next conversion in the
interrupt from the previous conversion rather than to just enable scan
mode.

This now actually reads all three sensors properly.