feat(heater-shaker): read NTC thermistors (#161)
This commit adds support for reading the onboard thermistors and the heater pad thermistors.

The way you convert from NTC resistance (which is temperature dependent)
into a temperature is by using a vendor-provided lookup table (since
otherwise there's lots of lovely ln and exp in there that you don't want
to deal with). This is (in this case) provided in a csv; we'd like it as
c++ source. So let's use cmake's ability to add custom targets and use
the output of those custom targets really easily to check in the csv and
generate c++ source and headers using a python script.

The new heater_hardware follows or perhaps starts or intensifies the
pattern in the motor where there needs to be a bridge, written by us,
between the c code with c expectations in the hal and the c++ code with
c++ expectations in the rest of the firmware. For instance, just because
we can turn off diagnostics about improper (for c++) use of volatile,
that doesn't mean we can actually use the macros that cause the warning;
we need to do that from C.

So, heater_hardware.c and .h provide a C interface to the hardware that
is usable by the c++ side.

This code brings up the ADC used by the thermistor pins. It's configured
so that the C++ code kicks off a three-channel conversion sequence, with
an interrupt firing after each conversion and eventually (when all three
channels have been converted) firing off a callback that should be
provided by the c++ side.

There's an annoying bit of statefulness in this code around remembering
which adc channel is currently being converted; the hardware doesn't do
it, the hal doesn't do it, so we have to do it. This also makes the
defensive coding in the conversion complete callback a little tricky
because even if we're not initialized yet (not that the callback should
fire before initialization) we need to make sure we don't get out of
sync with the ADC's internal state.

The thermistor conversion class ingests ADC readings directly (with
parametrization about the adc bit depth and the hardware parameters for
the thermistor circuit) and uses the autogenerated tables for the
thermistor and some other math to convert the values to C. It's also
almost monadic but I'm too much of a coward to go all the way. The
Result types get carried through and bring any errors in the
intermediates back to the caller rather than having to try and signal it
in-band by carving out like -1 or -inf as an error signal.

The numbers on the bias resistance in the test are slightly weird
because I forgot that number needed to be in kOhm rather than Ohm but
the numbers in the tests aree all arbitrary anyway, it's the boundary
checking around the std::find that needs to be tested.

Adds a message that will come from a hardware task and carry updated ADC
readings to the heater task. The heater task then converts the readings
to temperatures and handles any errors that result.

There needs to be a followup that handles things like the case where the
thermistor is broken and only gives bad readings; right now, we'll just
send an error every single time which isn't the most helpful thing in
the world.

As a sidenote, the temperature update messages going in provide a useful
but somewhat sideways way of specifying the control loop frequency. I
think it's ok that it's implicit since the actual sequence will be
specified by the value defined in this class.

For some reason the freertos call vTaskDelayUntil, which is the standard
way in an rtos to get stable timing in calls (you can't just use a
relative delay because the time you take actually doing stuff will add
up over time) isn't available by default. Now it is.


As with the motor task, we need another internal task to drive timing
and hardware-dependent stuff we don't want to sully the perfection of
the main c++ task. The task in this case does a bit less than in the
motor; it just serves to kick off a temperature update every so often.
This happens every so often rather than as fast as possible for two
reasons:
- The physical time scale we're talking about here is thermal, which
means slow. There's no point reading these ADCs in the megahertz (or
kilohertz, or frankly hertz) range because temperatures don't change
that fast; NTCs don't react that fast; and we have an antialiasing
filter in front of us anyway.
- There's no point having a million interrupts firing all the time if
we're not going to use the values.

So instead, we have this heater hardware task that runs (for now) at
10Hz (which can probably be lowered to more like 1 or 0.5Hz) and kicks
off an ADC conversion, which will cause 3 interrupts, one for each
channel, and send a message to wake the heater task on the third. That
means the main heater task will also run at 10Hz, just 3 adc conversions later.

Annoyingly, this micro's ADCs can be set up in chain conversion mode but
they only give you an interrupt after all the conversions are complete,
so if you're not either polling to catch it or using DMA you won't see
the first two. Of course, we didn't have it in that mode anyway before
that, but that's why the fix is to start the next conversion in the
interrupt from the previous conversion rather than to just enable scan
mode.