feat(heater-shaker): Add a heater policy
Like the motor task and the motor policy, the heater task needs a way to
proactively interact with the heater hardware. For now, this only has
the API to control the heater pad power good latching error
detector (and it's not actually implemented on the hardware side, and
that behavior isn't actually integrated anywhere; but due to the sheer
number of files touched, might as well make it its own commit.