feat(cmake-utils): add support to find STM32G4 BSP
For the ot3-firmware directory, we need to start supporting G4 nucleo boards. The file added here
will search github for the G4 Cube directory and add all the necessary files for the BSP. We did not
include the USB library as we will likely not be using USB capabilities for this board.

closes https://github.com/opentrons/ot3-firmware#6