feat(heater-shaker): Handle the heater power error latch (#165)
- feat(heater-shaker): Add a heater policy
Like the motor task and the motor policy, the heater task needs a way to
proactively interact with the heater hardware. For now, this only has
the API to control the heater pad power good latching error
detector (and it's not actually implemented on the hardware side, and
that behavior isn't actually integrated anywhere; but due to the sheer
number of files touched, might as well make it its own commit.

- feat(heater-shaker): handle heater error latch
The heater-shaker has some digital logic that ANDs all the various error
conditions (thermistor disconnected, thermistor short, thermistor
overtemp, etc) into each other and an SR latch that latches until reset
by the heater shaker. The software needs to handle this by making sure
the power is good before trying to control the heat pad.

This is implemented in this PR by
- Exposing the hardware API from C for sensing and resetting the error
latch
- Sending appropriate errors upstream
- Trying to proactively reset the latch when we detect in software that
an error condition has ended
- Trying to reset the latch if it's set when we try and set a temperature

Also make it possible to actually print out floats. I'm really glad newlib is committed to keeping code sizes low!!!!!