feat(heater-shaker): add bootloader handling
There's now a dfu command which will dispatch a message to the system
task; this task will then turn off peripherals and (not implemented yet)
jump to the bootloader after taking down the rest of the system. This
commit covers all the parts that link up the c++ side of things.

There's a minor hack in here with the new from_system bools in some
message that get us out of having to do full message origin tracking but
I think it's justified since it's literally for two messages,
immediately before the system goes down.