feat(heater-shaker): quick versioning
While we don't want full tag based versioning yet, it's nice to have
something. The heater-shaker now captures the current git
status (unfortunately, you'll need to do a make clean to update it -
this is mostly useful from CI) and put it in the build triggered by a
new gcode called version.