fixup: send back temperatures
While the motor task relies on querying the motor SDK internals to get
its value, the heater task requires the conversion interrupts to send it
messages about the state of the world. If we don't implement those in
the simulator, we'll never get a current temperature back. And those
messages are raw ADC readings, which we now need to generate from
temperatures. So,
- Add a way to go from temperature -> adc
- Add a way to get the setpoint (please don't use this in firmware)
- Send those temperature update messages if the setpoint changes