feat(heater-shaker): Add bootloader entry (#169)
* refactor(heater-shaker): ui task->system task (mv)

The heater/shaker won't actually have a ui really so there's no point
havinga ui task; and it has a couple more needs around features that
don't really fit super cleanly into other tasks like the soft power
button and bootloader entry. So let's make a system task. This is the
first step, moving the files around.

* refactor(heater-shaker): ui->system (renames)

Does the second part of the transition, renaming everything as necessary.

* feat(heater-shaker); add empty() to ack cache

A cache will be used in the system task as a proxy for a list of
remaining tasks before reboot, so we need to be able to check if it's empty.

* feat(heater-shaker): add bootloader handling

There's now a dfu command which will dispatch a message to the system
task; this task will then turn off peripherals and (not implemented yet)
jump to the bootloader after taking down the rest of the system. This
commit covers all the parts that link up the c++ side of things.

There's a minor hack in here with the new from_system bools in some
message that get us out of having to do full message origin tracking but
I think it's justified since it's literally for two messages,
immediately before the system goes down.

* add bootloader entry

Once we can deactivate the rest of the hardware, actually jump to the bootloader with some inline asm so the c compiler doesn't emit an !ldmia after we set the msp