fix(FindSTM32F303BSP): reverse driver compilation
The STM32F3 hal uses a pattern of having functions defined as __weak in
stm32f3xx_hal_subsystem.c and then non-weak overrides in
stm32f3xx_hal_subsystem_ex.c. When you compile to a static library as we
do, the object files compiled from these source files are packed in the
order they were passed to the compiler, which is alphabetical because we
use GLOB_RECURSE to list all the files. That means subsystem.c comes
before subsystem_ex.c.

That in turn means that when you link the static library to our
executable, the linker finds the weak definition first, says, "OK that's
good enough", and stops looking.

This fix reverses the alphabetical order, which means the subsystem_ex
files come before the subsystem files, which means the correct symbol
gets  linked, which you can verify by linking with
-Wl,-trace-symbol=some-overridden-symbol or by using objdump -t to list
everything in the STM32F303BSP_Drivers.a library and noting that
subsystem_ex files come before subsystem  files.