feat(heater-shaker): homing with solenoid (#178)
This PR adds the capability to home the plate, moving it slowly and engaging a solenoid to lock it in place.

That includes

    Bringing up the solenoid driver, which is an allegro h-bridge with current limit set by a DAC
    Adding a debug gcode, G28.D Sxxx where xxx is a current in integer milliamps (0 means disengage), to test the solenoid
    Adding a home gcode, G28, that starts the homing process
    Adding a home-process state machine to the motor task

The home-process state machine is a little involved because it wants to not block the motor task while it's going. This is done by having the motor task send a message back to itself every so often to check the homing status when it's in that state.