update interrupt priorities to speed up mclib
The library really wants to run its controller at 30khz. It also really
wants to race timeers to do this. Things need to not be delayed. So we
lowered the priority of less latency-sensitive stuff like usb, and
raised the priority of the motor control library stuff above where the
FreeRTOS priority limits live (based on
configLIBRARY_MAX_SYSCALL_INTERRUPT_PRIORITY). This means we  can't use
freeretos stuff from inside the core mclib, but we weren't going to anyway.