feat(heater-shaker): discrete-time pid
When you analyze a PID controller that is magical (or analog), you can
use continuous math. When you analyze a PID controller that is code, you
can't. The PID implementation is now done with the euler approximation
discretization approach (t ~= kT, integer k 0<k<N, T=sampling time,
dx(t)/dt ~= (x(k)-x(k-1))/T) which yields more accurate results.

Also, while we previously had a windup limiter for our integrator,
another (possibly better, possibly just complementary) approach is to
recognize the situations in which the i-term winds up and reset it after
that situation is exited. We have an armable integrator reset, which you
can arm to look for the next time we cross the setpoint, and reset the
integrator there. This can be more responsive than a windup limiter.