chore(heater-shaker): heater tuning v1 (#190)
* fixup: fix heater power setting

Setting heater power felt very silly what with the turning the timer off
and back on every time. Instead, let's do it one time, and then use the
magic macro that just sets the compare  register.

* feat(heater-shaker): discrete-time pid

When you analyze a PID controller that is magical (or analog), you can
use continuous math. When you analyze a PID controller that is code, you
can't. The PID implementation is now done with the euler approximation
discretization approach (t ~= kT, integer k 0<k<N, T=sampling time,
dx(t)/dt ~= (x(k)-x(k-1))/T) which yields more accurate results.

Also, while we previously had a windup limiter for our integrator,
another (possibly better, possibly just complementary) approach is to
recognize the situations in which the i-term winds up and reset it after
that situation is exited. We have an armable integrator reset, which you
can arm to look for the next time we cross the setpoint, and reset the
integrator there. This can be more responsive than a windup limiter.

* policy: clamp inputs instead of ignoring

Because come on.

* heater-shaker: scripts: modularize

Pull a bunch of stuff from speed_stability to new file test_utils for reuse

* heater-shaker: scripts: thermal analysis scripts

heater_profiling generates a step-response and saves and plots thee
data; heater-stability works like speed-stability, sets a setpoint and
records temperature.

* different heater pid constants

Bit better tuned. May neeed to up the  damping factor.

* fix heater-power-test

We were turning off the heater right after setting the power

* raise the adc sample time

because of the design of the adc and the electrical characteristics of
the thermistor circuit we need to give the sampling cap a very long time
to charge to work properly.