fix(heater-shaker): set usb class in device desc (#209)
The default configuration for the USB descriptors defers the device
class to each interface. USB CDC/ACM has two interfaces - one for
control (the ACM, class 2 subclass 2) and one for data (the CDC, class a
subclass 0). If you don't specify a sub/class in the _device_
descriptor, the spec says hosts look at each interface separately, and
windows serial driver binding is looking for class 2 subclass 2, so it
binds the ACM interface but not the CDC interface, and then fails the
device.

By specifying class and subclass somewhat redundantly in the top level
device descriptor, windows binds the composite device to the cdc/acm
drivers and all is good.