chore(thermocycler-refresh): Fix Lint configuration (#222)
Fixed `-lint` targets for `common` and `thermocycler-refresh` to actually load the config file in our root folder. Also enabled checking for implicit switchcase fallthrough, because I couldn't get it to configure in the .clang-tidy file and it is really helpful for avoiding some common bugs. Also fixed up the thermocycler-refresh code to happily pass the linting check.