refactor: add helpers for heater task checks
A bunch of heater task tests had the same sets of like 3 lines around
checking for responses in the host comms task queue or sending the
heater task a message and then running it once. By pulling these into
helper methods we can cut down on the verbosity and on the amount of
stuff you have to remember to do when writing these tests.