feat: add error handling to thermal plate
The error handling works like this:

- We will save an "interrupted-by" error that persists until the next
GetErrorState message, even if the error that caused it originally goes
away. This is important so that we don't miss transient errors that
cancelled any ongoing work
- We get an ongoing error state in preference to an interrupted-by
- When we clear the error state, it's not really committed (i.e., the
LEDs don't change) until the next temperature conversion, to avoid any
weird flickering. Not that you'd see that in practice for now, since
it's a test command.
- When we set the manual error, it may get overridden by a subsequent
real error.